requirements.txt
opencv==4.6.0
streamlit==1.27.0
pillow==9.4.0